test: drop the cuda_event query tests
Testing that cudaEventQuery returns cudaSuccess after a sync, cudaErrorNotReady
while in flight, and an error for a bad handle characterizes the CUDA runtime,
not cuCascade. The cuCascade code underneath is a three-branch status mapping
and two one-line forwards, and its success and not-ready arms are already
exercised through the reader-event tests in test_data_batch.cpp.

Removing the file leaves the stream_gate helpers with a single consumer again,
so move them back into test_data_batch.cpp and drop the shared header.
